Ruby worker hooks breakdown to modules, worker loading external files
