+++ test/fixtures/ruby-hooks/hookfile2.rb
@@ -1,0 +1,5 @@
+include DreddWorker::Hooks
+
+after_all do |transactions|
+  puts 'ruby hooks second file'
+end
